requirements.txt
python-telegram-bot==20.7
gspread==6.0.2
google-auth==2.27.0
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
anthropic==0.40.0
# mcp — only needed for mock_server.py (Claude Code CLI), requires Python 3.10+
# Install separately: pip install mcp
